Create Offer forms

+++ OffersManagementSystem.Application/IRepositories/IOfferRepository.cs
@@ -10,7 +10,7 @@ namespace OffersManagementSystem.Application.IRepositories
     public interface IOfferRepository
     {
         Task<IEnumerable<Offer>> GetAllOffersAsync(string? serial, string? offerAddress, DateOnly? fromDate, DateOnly? toDate, bool? active, decimal? totalFrom, decimal? totalTo);
-        Task<Offer> GetOfferByIdAsync(int id);
+        Task<Offer> GetOfferByIdAsync(long id);
         Task<int> CreateOfferAsync(Offer offer);
         Task<int> UpdateOfferAsync(Offer offer);
         Task<int> DeleteOfferAsync(int id);
